[layer] add Mean & MeanReduce layer
added MeanLayer (It contains MeanReduce operation, too)

if the "axis" peroperty is not set, then it performs a mean operation
averaging the entire tensor.

if the axis peroperty is specified, it performs a MeanReduce operation
that averages along the specified axis. (However, in NNTrainer, since
the axis does not actually disappear, the length of the axis is changed to 1).

Signed-off-by: Seungbaek Hong <sb92.hong@samsung.com>