[ bugfix ] Proper initialization of struct
The previous code
`S = { 0 };`
failed because it attempted to initialize the first member of the struct
`std::vector<float> observation` with the integer 0, which is not a valid operation.

The fix `S = {};`
uses Value Initialization for the entire struct. This tells the compiler to initialize every member with its default empty/zero state

**Self evaluation:**
1. Build test:     [X]Passed [ ]Failed [ ]Skipped
2. Run test:     [X]Passed [ ]Failed [ ]Skipped

Signed-off-by: skykongkong8 <ss.kong@samsung.com>